Now tracking all stats for SJF non preemptive

+++ Scheduler/Models/SchedulerStats.cs
@@ -11,6 +11,6 @@ namespace Scheduler.Models
         public double AverageTurnAroundTime { get; set; }
 
         public Dictionary<string, int> ProcessWaitTimes { get; set; }
-        public Dictionary<string, int> ProcessTurnAroundTimes { get; set; }
+        public Dictionary<string, double> ProcessTurnAroundTimes { get; set; }
     }
 }
